Refactor db initialize part
