+++ app/backend/routers/streamer_r.go
@@ -1,11 +1,12 @@
 package routers
 
 import (
+	"database/sql"
 	"github.com/gin-gonic/gin"
 	"tchatong.info/controllers"
 )
 
-func SetStreamerRouter(router *gin.RouterGroup) {
-	router.GET("", controllers.GetAllStreamer)
-	router.GET("/:streamerId", controllers.GetOneStreamer)
+func SetStreamerRouter(router *gin.RouterGroup, db *sql.DB) {
+	router.GET("", func(c *gin.Context) { controllers.GetAllStreamer(c, db) })
+	router.GET("/:streamerId", func(c *gin.Context) { controllers.GetOneStreamer(c, db) })
 }
